+++ components/molecules/calendar/calendar.tsx
@@ -5,7 +5,7 @@ import { ChevronLeftIcon, ChevronRightIcon } from "@radix-ui/react-icons";
 import { DayPicker } from "react-day-picker";
 
 import { cn } from "@/lib/utils";
-import { buttonVariants } from "../button";
+import { buttonVariants } from "../../atoms/button";
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>;
